添加了反序列化自定义属性名称
